requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.2.0
openpyxl>=3.1.0
pyyaml>=6.0.0
click>=8.1.0
tqdm>=4.66.0
structlog>=24.1.0
urllib3>=2.0.0

